+++ provider/provider-settings/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.0.7 (2022-05-02)
+
+**Note:** Version bump only for package @gemunion/provider-settings
+
+
+
+
+
 ## 2.0.6 (2022-04-28)
 
 **Note:** Version bump only for package @gemunion/provider-settings