+++ provider/provider-user/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 3.3.2 (2022-05-02)
+
+**Note:** Version bump only for package @gemunion/provider-user
+
+
+
+
+
 ## 3.3.1 (2022-04-28)
 
 **Note:** Version bump only for package @gemunion/provider-user