+++ react/geetest/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.3.22 (2022-05-02)
+
+**Note:** Version bump only for package @gemunion/geetest
+
+
+
+
+
 ## [1.3.21](https://github.com/gemunion/common-packages/compare/@gemunion/geetest@1.3.20...@gemunion/geetest@1.3.21) (2022-04-15)