+++ react/hooks/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.1.0 (2022-05-02)
+
+
+### Bug Fixes
+
+* useCollection ([733751d](https://github.com/gemunion/mui-packages/commit/733751dd102e1b6f06201f27e26a1fe51d1f013e))
+
+
+### Features
+
+* useCollection ([9e39272](https://github.com/gemunion/mui-packages/commit/9e392725178f7769838b2d77b53d646edbaaabcc))
+
+
+
+
+
 ## 0.0.2 (2022-03-05)